% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{gtm_container_version}
\alias{gtm_container_version}
\title{GTM version list}
\usage{
gtm_container_version(accountId, containerId)
}
\arguments{
\item{accountId}{Add your GTM account ID}

\item{containerId}{Add your GTM container ID}
}
\description{
Downloads all GTM versions for a defined container ID
}
