% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{gtm_create_container}
\alias{gtm_create_container}
\title{GTM create a container}
\usage{
gtm_create_container(
  accountId,
  containerName,
  containerType,
  containerDomain = NULL
)
}
\arguments{
\item{accountId}{Add your GTM account ID}

\item{containerName}{Name of the container you want to create}

\item{containerType}{Type of container (web or amp)}

\item{containerDomain}{Domainname related to the container (optional)}
}
\description{
Create a GTM container
}
