% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{gtm_create_folder}
\alias{gtm_create_folder}
\title{GTM create a folder}
\usage{
gtm_create_folder(accountId, containerId, workspaceId, folderName)
}
\arguments{
\item{accountId}{Add your GTM account ID}

\item{containerId}{Add your GTM container ID}

\item{workspaceId}{Add the workspace ID from your container}

\item{folderName}{Name of folder you would like to create}
}
\description{
Create a folder in GTM container
}
