% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{gtm_environment_list}
\alias{gtm_environment_list}
\title{GTM Environment list}
\usage{
gtm_environment_list(accountId, containerId)
}
\arguments{
\item{accountId}{Add your GTM account ID}

\item{containerId}{Add your GTM container ID}
}
\description{
Downloads all GTM environments for a defined container ID
}
