% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{gtm_get_account}
\alias{gtm_get_account}
\title{Get individual GTM account detail}
\usage{
gtm_get_account(accountId)
}
\arguments{
\item{accountId}{accountId to request}
}
\value{
account data list of $accountId, $name, $shareDate, $fingerprint
}
\description{
Get individual GTM account detail
}
