% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{gtm_get_container}
\alias{gtm_get_container}
\title{Get individual GTM account detail}
\usage{
gtm_get_container(accountId, containerId)
}
\arguments{
\item{accountId}{accountId to request}

\item{containerId}{GTM containerId}
}
\value{
container info
}
\description{
Get individual GTM account detail
}

