% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{gtm_get_current_version}
\alias{gtm_get_current_version}
\title{GTM current Version}
\usage{
gtm_get_current_version(accountId, containerId)
}
\arguments{
\item{accountId}{Add your GTM account ID}

\item{containerId}{Add your GTM container ID}
}
\description{
Downloads the most recent GTM version for a defined container ID
}
