% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{gtm_get_variables}
\alias{gtm_get_variables}
\title{Get GTM variables}
\usage{
gtm_get_variables(accountId, containerId, variableId)
}
\arguments{
\item{accountId}{GTM account ID}

\item{containerId}{GTM Container ID}

\item{variableID}{The GTM variable ID}
}
\value{
Dataframe of accountId, name
}
\description{
Get GTM variables
}
\seealso{
\href{https://developers.google.com/tag-manager/api/v1/reference/accounts/containers/variables/get}{Google Docs}
}
