% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{gtm_list_builtin}
\alias{gtm_list_builtin}
\title{GTM builtin Variables list}
\usage{
gtm_list_builtin(accountId, containerId)
}
\arguments{
\item{accountId}{Add your GTM account ID}

\item{containerId}{Add your GTM container ID}
}
\description{
Downloads all GTM Builtin Variables (name and type) that are enabled in the current workspace, container
}
