% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{gtm_list_containers}
\alias{gtm_list_containers}
\title{GTM container list}
\usage{
gtm_list_containers(accountId)
}
\arguments{
\item{accountId}{Add your GTM account ID to get the list of containers}
}
\description{
Downloads all GTM containers for a defined account ID
}
