% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{gtm_list_environments}
\alias{gtm_list_environments}
\title{GTM Environment list}
\usage{
gtm_list_environments(accountId, containerId)
}
\arguments{
\item{accountId}{Add your GTM account ID}

\item{containerId}{Add your GTM container ID}
}
\description{
Downloads all GTM environments for a defined container ID
}
