% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{gtm_list_folders}
\alias{gtm_list_folders}
\title{GTM folders list}
\usage{
gtm_list_folders(accountId, containerId)
}
\arguments{
\item{accountId}{Add your GTM account ID}

\item{containerId}{Add your GTM container ID}
}
\description{
Downloads all folders available in the container
}
