% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{gtm_list_tags}
\alias{gtm_list_tags}
\title{GTM tags list}
\usage{
gtm_list_tags(accountId, containerId)
}
\arguments{
\item{accountId}{Add your GTM account ID}

\item{containerId}{Add your GTM container ID}
}
\description{
Downloads all GTM tags that are enabled in the current workspace, container
}
