% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{gtm_list_templates}
\alias{gtm_list_templates}
\title{GTM template list}
\usage{
gtm_list_templates(accountId, containerId, workspaceId)
}
\arguments{
\item{accountId}{Add your GTM account ID}

\item{containerId}{Add your GTM container ID}

\item{workspaceId}{Add the workspace ID from your container}
}
\description{
Downloads all templates available in the container
}
