% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triggers.R
\name{gtm_list_triggers}
\alias{gtm_list_triggers}
\title{GTM triggers list}
\usage{
gtm_list_triggers(accountId, containerId)
}
\arguments{
\item{accountId}{Add your GTM account ID}

\item{containerId}{Add your GTM container ID}
}
\description{
Downloads all GTM triggers that are enabled in the current workspace, container
}
