% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{gtm_list_users}
\alias{gtm_list_users}
\title{GTM users list}
\usage{
gtm_list_users(accountId)
}
\arguments{
\item{accountId}{Add your GTM account ID}
}
\description{
Downloads all users that have access to the account
}
