% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{gtm_list_variables}
\alias{gtm_list_variables}
\title{List GTM variables}
\usage{
gtm_list_variables(accountId, containerId)
}
\arguments{
\item{accountId}{GTM account ID}

\item{containerId}{GTM Container ID}
}
\value{
Dataframe of accountId, name
}
\description{
List GTM variables
}

