% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{gtm_list_workspaces}
\alias{gtm_list_workspaces}
\title{GTM workspace list}
\usage{
gtm_list_workspaces(accountId, containerId)
}
\arguments{
\item{accountId}{Add your GTM account ID}

\item{containerId}{Add your GTM container ID}
}
\description{
Downloads all GTM workspaces for a defined container ID
}
