% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\name{gtm_list_zones}
\alias{gtm_list_zones}
\title{GTM Zones list}
\usage{
gtm_list_zones(accountId, containerId, workspaceId)
}
\arguments{
\item{accountId}{Add your GTM account ID}

\item{containerId}{Add your GTM container ID}

\item{workspaceId}{Add the workspace ID from your container}
}
\description{
Downloads all GTM workspaces for a defined container ID
}
