% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{gtm_workspace_id}
\alias{gtm_workspace_id}
\title{GTM current workspace ID}
\usage{
gtm_workspace_id(accountId, containerId)
}
\arguments{
\item{accountId}{Add your GTM account ID}

\item{containerId}{Add your GTM container ID}
}
\description{
Downloads current workspace ID
}
